# openapi.yaml
openapi: 3.0.0
info:
  title: Group Management API
  description: API for managing groups using Hexagonal Architecture
  version: 1.0.0
  
servers:
  - url: http://localhost:3000
    description: Local development server

paths:
  /groups:
    get:
      summary: List all groups
      description: Retrieves a list of all groups
      responses:
        '200':
          description: Successfully retrieved groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
    post:
      summary: Create a new group
      description: Creates a new group with the given name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
              required:
                - name
      responses:
        '201':
          description: Group created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid

  /groups/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: The group ID
    
    get:
      summary: Get a specific group
      description: Retrieves details of a specific group
      responses:
        '200':
          description: Successfully retrieved group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '404':
          description: Group not found
    
    patch:
      summary: Update a group
      description: Updates an existing group's information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '204':
          description: Group updated successfully
        '404':
          description: Group not found
    
    delete:
      summary: Delete a group
      description: Deletes an existing group
      responses:
        '204':
          description: Group deleted successfully
        '404':
          description: Group not found

components:
  schemas:
    Group:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt 